add animation class, improvements
- added animation class (both vertical and horizontal sprite sheet stacking supported)
- address some of the warnings (implicit conversion from int to unsigned long int)
- added game mode page

+++ source_code/readDatasetCSV.cpp
@@ -49,7 +49,7 @@ bool readDatasetCSV(std::string filename, trieNode*& pRoot)
         {
             std::cout << "[DEBUG] end of vec\n";
             insert(pRoot, previousWord, definitionVec);
-            for (int i = 0; i < definitionVec.size(); ++i)
+            for (int i = 0; i < (int)definitionVec.size(); ++i)
             {
                 // std::cout << "[DEBUG]" << previousWord << " ";
                 // std::cout << definitionVec[i].first << " ";
